Fix presubmit errors
